Split sync delta planners
