+++ service-worker.js
@@ -149,6 +149,10 @@ const APP_SHELL = [
   '/js/sync-schema.js',
   '/js/sync-delta.js',
   '/js/sync-delta-planners.js',
+  '/js/sync-delta-planner-context.js',
+  '/js/sync-delta-array-planner.js',
+  '/js/sync-delta-map-planner.js',
+  '/js/sync-delta-scalar-planner.js',
   '/js/sync-delta-snapshot.js',
   '/js/sync-delta-merge.js',
   '/js/sync-delta-merge-shapes.js',
